requirements.txt
streamlit
Pillow
numpy
opencv-python-headless
pytesseract
# tesseract-ocr
# click==8.1.3